requirements.txt
# ── Combined CTF Platform + Evaluation API ───────────────────────────────────

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0

# Templates
jinja2>=3.1.0

# Auth / sessions
itsdangerous>=2.1.0
starlette>=0.36.0

# Database
sqlalchemy>=2.0.0

# Password hashing
werkzeug>=3.0.0

# HTTP (challenge asset downloads)
requests>=2.28.0

# Data & ML (Challenges 1-5 evaluation)
pandas>=2.0.0
numpy>=1.23.0
scikit-learn>=1.3.0
joblib>=1.2.0
